LayoutTests/fast/multicol/newmulticol/balance10.html

<!DOCTYPE html>
<html>
  <head>
    <title>auto-height multicol</title>
    <script src="../../../resources/js-test-pre.js"></script>
    <script>
      if (window.testRunner)
        testRunner.dumpAsText();
      if (window.internals)
        internals.settings.setRegionBasedColumnsEnabled(true);
    </script>
  </head>
  <body>
    <p>There should be two identical boxes below.</p>
    <div id="mc" style="-webkit-columns:3; columns:3; orphans:1; widows:1; line-height:50px; margin:1em; float:left; width:18em; color:olive; background:olive;">
      <span>line</span><br>
      <span>line</span><br>
      <span>line</span><br>
      <span>line</span><br>
      <span>line</span><br>
      <span>line</span><br>
      <span>line</span><br>
      <span style="line-height:150px;">line</span><br>
      <span>line</span><br>
      <span>line</span><br>
    </div>
    <div id="ref" style="-webkit-columns:3; columns:3; orphans:1; widows:1; line-height:50px; margin:1em; float:left; width:18em; height:250px; color:olive; background:olive;">
      <span>line</span><br>
      <span>line</span><br>
      <span>line</span><br>
      <span>line</span><br>
      <span>line</span><br>
      <span>line</span><br>
      <span>line</span><br>
      <span style="line-height:150px;">line</span><br>
      <span>line</span><br>
      <span>line</span><br>
    </div>
    <script>
      shouldBe("mc.offsetHeight", "ref.offsetHeight");
    </script>
    <script src="../../../resources/js-test-post.js"></script>
  </body>
</html>
